+++ packages/runtime/src/components/Bus/Enabled/index.tsx
@@ -1,3 +1,11 @@
+/*
+ * @Description: 
+ * @Version: 1.0
+ * @Autor: yangyuhang
+ * @Date: 2022-06-20 10:58:33
+ * @LastEditors: yangyuhang
+ * @LastEditTime: 2023-04-12 18:20:25
+ */
 import React, { useState, useMemo, useEffect } from 'react';
 import { Switch } from 'antd';
 import { BsTableComponentProps } from '../../Base/BsTable';
@@ -10,6 +18,7 @@ type EnabledProps = BsTableComponentProps & {
   request: (params: any) => Promise<any>; // 请求数据的远程方法
   rowKeyName?: string;
   warning?: string;
+  openWarning?: string;
   enabledName?: string;
   funcode?: string;
   disabled?: boolean;
@@ -24,6 +33,7 @@ const Enabled: React.FC<EnabledProps> = (props) => {
     value,
     rowData,
     warning = '您是否确定禁用?',
+    openWarning,
     enabledName = 'enabled',
     disabled,
     callback,
@@ -58,14 +68,33 @@ const Enabled: React.FC<EnabledProps> = (props) => {
         },
       });
     } else {
-      run({
-        [rowKeyName]: rowData[rowKeyName],
-        [enabledName]: checked,
-      }).then(() => {
-        setState(checked);
-        rowData[enabledName] = checked;
-        callback?.(true);
-      });
+      if (openWarning) {
+        CModal.confirm({
+          title: `${openWarning}`,
+          okText: '确定',
+          cancelText: '取消',
+          onOk: () => {
+            return run({
+              [rowKeyName]: rowData[rowKeyName],
+              [enabledName]: checked,
+            }).then((data) => {
+              setState(checked);
+              rowData[enabledName] = checked;
+              callback?.(true);
+              return true;
+            });
+          },
+        });
+      } else {
+        run({
+          [rowKeyName]: rowData[rowKeyName],
+          [enabledName]: checked,
+        }).then(() => {
+          setState(checked);
+          rowData[enabledName] = checked;
+          callback?.(true);
+        });
+      }
     }
   };