+++ packages/runtime/src/components/Bus/MallGoods/Catalog/CatalogTreeSelect.tsx
@@ -1,3 +1,11 @@
+/*
+ * @Description: 
+ * @Version: 1.0
+ * @Autor: yangyuhang
+ * @Date: 2023-01-05 10:01:22
+ * @LastEditors: yangyuhang
+ * @LastEditTime: 2023-07-05 11:02:01
+ */
 import React from 'react';
 import { ScTreeSelect } from '@scboson/sc-element';
 
@@ -13,7 +21,7 @@ import { uesRequest } from '../../../../utils/api';
  */
 const CatalogTreeSelect: React.FC<
   ScTreeSelectProps & { selectLeaf: boolean }
-> = ({ params, selectLeaf = true, autoload = true, ...restProps }) => {
+> = ({ params = {}, selectLeaf = true, autoload = true, ...restProps }) => {
   const { run, loading } = uesRequest('mallgoods_catalog', 'treeList');
 
   return (
@@ -35,7 +43,7 @@ const CatalogTreeSelect: React.FC<
       treeDataSimpleMode
       autoload={true}
       request={run}
-      params={{}}
+      params={params}
       textField="catalogName"
       valueField="catalogId"
       {...restProps}