+++ packages/runtime/src/components/WithComponent/interface.ts
@@ -9,6 +9,8 @@ import type {
 import type { MutableRefObject } from 'react';
 
 export type WithTableProps = {
+  initialValues?: any;
+  autoSubmitFiled?: any[];
   extraColumns?: ProColumn[];
   extraQueryColumns?: FormSearchItem[];
   selectionType?: 'checkbox' | 'radio';